changed type of rate field to DECIMAL in product model and created migration for it

+++ src/sequelize/models/product.model.js
@@ -40,7 +40,7 @@ module.exports = (sequelize) => {
                 onUpdate: 'CASCADE'
             },
             rate: {
-                type: DataTypes.INTEGER,
+                type: DataTypes.DECIMAL(2, 1),
                 defaultValue: 5,
                 validate: {
                     min: 0,
